Fix Android compilation errors and build issues
- Fix serializer issues in SecureDataRepository.kt by adding missing import and using modern Kotlin serialization syntax
- Resolve type mismatches between domain and data models by standardizing on data models in repository layer
- Add missing getTotalEntityCount() member function to data AppData class
- Fix platform declaration clashes in extension functions by adding @JvmName annotations
- Add explicit type annotations in migration extensions to resolve type inference issues
- Update BUILD_ISSUE.md to track resolved compilation errors

All 44+ compilation errors have been resolved. The Android project should now build successfully.

+++ frontend/app/src/main/java/com/pocketagent/domain/models/AppData.kt
@@ -110,7 +110,7 @@ data class AppData(
      * Gets total storage size estimate in bytes.
      */
     fun getEstimatedStorageSize(): Long {
-        val baseSize = toString().toByteArray().size
+        val baseSize = toString().toByteArray().size.toLong()
         val messagesSize = messages.values.flatten().sumOf { it.getSize() }
         return baseSize + messagesSize
     }
